UI beginnings for publication flow
