+++ theme/templates/resource-landing-page/modals.html
@@ -484,7 +484,7 @@
             </div>
             <div class="modal-footer">
                 <button type="button" class="btn btn-default" data-dismiss="modal">Cancel</button>
-                <a type="button" id="publish-btn-2" class="btn btn-danger" href="/hsapi/_internal/{{ cm.short_id }}/publish/">Publish</a>
+                <a type="button" id="publish-btn-2" class="btn btn-danger" href="#">Publish</a>
             </div>
         </div>
     </div>
